+++ NEWS.md
@@ -1,5 +1,6 @@
 # Release Notes
 ## Version `v26.5.5` (in progress)
+* Use sessionStorage to ensure `isConsortium` url param is respected after form submit to allow flags to persist if errors are returned. (KEYCLOAK-51).
 
 ## Version `v26.5.4` (14.05.2026)
 * Use Keycloak supported `jdbc-ping` cache discovery instead of custom JDBC_PING2 XML while preserving authorization cache size tuning via `KC_CACHE_EMBEDDED_AUTHORIZATION_MAX_COUNT`; offline session cache limits can be tuned via `KC_CACHE_EMBEDDED_OFFLINE_SESSIONS_MAX_COUNT` and `KC_CACHE_EMBEDDED_OFFLINE_CLIENT_SESSIONS_MAX_COUNT` (KEYCLOAK-111)